s2 - server refactoring. Groups
