+++ server/controllers/controllers.js
@@ -1,4 +1,4 @@
-class ControllerConstr {
+class Controllers {
 	constructor(service) {
 		this.service = service;
 	}
@@ -13,29 +13,9 @@ class ControllerConstr {
 		}
 	}
 
-	async createWithFile(req, res) {
-		try {
-			const data = await this.service.createWithFile(req.body, req.files, "File");
-			res.json(data);
-		}
-		catch (error) {
-			res.status(500).json(error.message);
-		}
-	}
-
 	async getAll(req, res) {
 		try {
 			const data = await this.service.getAll();
-			res.json(data);
-		}
-		catch (error) {
-			res.status(500).json(error.message);
-		}
-	}
-
-	async dynamicLoading(req, res) {
-		try {
-			const data = await this.service.dynamicLoading(req);
 			res.json(data);
 		}
 		catch (error) {
@@ -75,17 +55,6 @@ class ControllerConstr {
 			res.status(500).json(error.message);
 		}
 	}
-
-	async deleteWithFile(req, res) {
-		try {
-			const {id} = req.params;
-			const data = await this.service.deleteWithFile(id);
-			res.json(data);
-		}
-		catch (error) {
-			res.status(500).json(error.message);
-		}
-	}
 }
 
-export default ControllerConstr;
+export default Controllers;
