+++ server/controllers/groups-controller.js
@@ -1,0 +1,6 @@
+import groupsService from "../services/groups-service.js";
+import Controllers from "./controllers.js";
+
+class GroupsController extends Controllers {}
+
+export default new GroupsController(groupsService);
