+++ server/index.js
@@ -3,11 +3,11 @@ import express from "express";
 import fileUpload from "express-fileupload";
 import mongoose from "mongoose";
 
-import albumRouter from "./routers/album-router.js";
-import filesRouter from "./routers/files-router.js";
+/* import albumRouter from "./routers/album-router.js";
+import filesRouter from "./routers/files-router.js"; */
 import groupRouter from "./routers/group-router.js";
-import memberRouter from "./routers/member-router.js";
-import styleRouter from "./routers/style-router.js";
+/* import memberRouter from "./routers/member-router.js";
+import styleRouter from "./routers/style-router.js"; */
 
 
 const PORT = 5000;
@@ -20,10 +20,10 @@ app.use(fileUpload({}));
 app.use(cors());
 
 app.use("/api", groupRouter);
-app.use("/api", memberRouter);
+/* app.use("/api", memberRouter);
 app.use("/api", albumRouter);
 app.use("/api", styleRouter);
-app.use("/api", filesRouter);
+app.use("/api", filesRouter); */
 
 async function startApp() {
 	try {
