+++ server/routers/group-router.js
@@ -1,18 +1,13 @@
 import Router from "express";
 
-import ControllerConstr from "../controllers/controller-constr.js";
-import getGroupsController from "../controllers/get-groups-controller.js";
-import Group from "../models/group.js";
-import ServicesConstr from "../services/services-constr.js";
+import groupsController from "../controllers/groups-controller.js";
 
-const groupService = new ServicesConstr(Group);
-const groupController = new ControllerConstr(groupService);
 const groupRouter = new Router();
 
-groupRouter.post("/groups", (req, res) => groupController.create(req, res));
-groupRouter.get("/groups", (req, res) => getGroupsController.getAll(req, res));
-groupRouter.get("/groups/:id", (req, res) => groupController.getOne(req, res));
-groupRouter.put("/groups/:id", (req, res) => groupController.update(req, res));
-groupRouter.delete("/groups/:id", (req, res) => groupController.delete(req, res));
+groupRouter.post("/groups", (req, res) => groupsController.create(req, res));
+groupRouter.get("/groups", (req, res) => groupsController.getAll(req, res));
+groupRouter.get("/groups/:id", (req, res) => groupsController.getOne(req, res));
+groupRouter.put("/groups/:id", (req, res) => groupsController.update(req, res));
+groupRouter.delete("/groups/:id", (req, res) => groupsController.delete(req, res));
 
 export default groupRouter;
