+++ server/services/groups-service.js
@@ -1,0 +1,6 @@
+import Group from "../models/group.js";
+import Services from "./services.js";
+
+class GroupsService extends Services {}
+
+export default new GroupsService(Group);
