+++ sources/views/groups/index.js
@@ -80,7 +80,7 @@ export default class Groups extends JetView {
 	}
 
 	init() {
-		groupsDB.load(groupsURL);
+		/* groupsDB.load(groupsURL); */
 		this._popup = this.ui(Popup);
 		const table = this.$$("table");
